// types.ts
export type Role = 'admin' | 'reception' | 'barber' | 'display' | 'cashier';
// ...
export interface User {
  id: string;
  name: string;
  username: string;
  role: Role;
  pin: string;
  active?: boolean;
  branchId?: string;
  canDoPos?: boolean;
}